requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
bitarray==3.0.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
ckzg==2.0.1
colorama==0.4.6
constantly==23.10.4
cors==1.0.1
cryptography==44.0.0
cytoolz==1.0.1
daphne==4.1.2
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.1.0
eth-utils==5.2.0
eth_abi==5.2.0
Faker==38.0.0
filelock==3.17.0
frozenlist==1.5.0
future==1.0.0
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
hexbytes==1.3.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
logging==0.4.9.6
msgpack==1.1.0
multidict==6.1.0
packaging==24.2
parsimonious==0.10.0
pillow==12.0.0
propcache==0.2.1
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyOpenSSL==25.0.0
PySocks==1.7.1
python-dotenv==1.2.1
pyunormalize==16.0.0
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rlp==4.1.0
service-identity==24.2.0
setuptools==75.8.0
sqlparse==0.5.3
tldextract==5.1.3
toolz==1.0.0
Twisted==24.11.0
txaio==23.1.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
web3==7.8.0
websockets==13.1
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
